pyproject.toml
[tool.poetry]
name = "threat-analysis-agents"
version = "0.1.0"
description = "网络安全分析调度智能体。判断用户请求是 domain, ip, sample, 还是 vulnerability 分析，并传递给对应智能体进行分析"
authors = ["kenshin <kenshin.acs@gmail.com>"]
readme = "README.md"
packages = [{include = "ta_agents"}]

[tool.poetry.dependencies]
python = "^3.10"
langfuse = "^2.60.4"
nest-asyncio = "^1.6.0"
pydantic-ai = {extras = ["logfire"], version = "^0.1.10"}
openai-agents = "^0.0.14"
aiohttp = "^3.11.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
